+++ contracts/dkg/SkaleDkgResponse.sol
@@ -87,9 +87,9 @@ library SkaleDkgResponse {
         bytes32[2] memory publicKey = INodes(contractManager.getContract("Nodes")).getNodePublicKey(
             complaints[schainHash].fromNodeToComplaint
         );
-        uint256 pkX = uint(publicKey[0]);
+        uint256 pkX = uint256(publicKey[0]);
 
-        (pkX, ) = IECDH(contractManager.getContract("ECDH")).deriveKey(secretNumber, pkX, uint(publicKey[1]));
+        (pkX, ) = IECDH(contractManager.getContract("ECDH")).deriveKey(secretNumber, pkX, uint256(publicKey[1]));
         bytes32 key = bytes32(pkX);
 
         // Decrypt secret key contribution
